- Cria app reservas, model reserva, ativa o admin

+++ anuncios/models.py
@@ -10,16 +10,16 @@ class Anuncio(models.Model):
 		HOTEL_URBANO = 'HOTEL_URBANO', 'Hotel Urbano'
 	imovel = models.ForeignKey(Imovel, on_delete=models.CASCADE)
 	plataforma = models.CharField(max_length=30, choices=Plataforma.choices)
-	taxa_plataforma = models.FloatField()
+	taxa_plataforma = models.DecimalField(decimal_places=2, max_digits=6)
 	dt_criacao = models.DateTimeField(auto_now_add=True)
 	dt_atualizacao = models.DateTimeField(auto_now=True)
 
 	def save(self, *args, **kwargs):
 		taxas_dict = {
-			'AIRBNB': 50,
-			'BOOKING': 60,
-			'TRIVAGO': 70,
-			'HOTEL_URBANO': 80,
+			'AIRBNB': 50.00,
+			'BOOKING': 60.00,
+			'TRIVAGO': 70.00,
+			'HOTEL_URBANO': 80.00,
 		}
 		self.taxa_plataforma = taxas_dict[self.plataforma]
 		super().save()
